Using Absolute Positioning & Floats
Demonstrate your proficiency with absolute positioning and floats. You have three layout tasks: Use absolute positioning to place the <div> with the banner—title class on top of the banner image, use a float to move the image within the highlights—media element to the left of the text, and construct the mosaic of paintings.